refatorando o model de evento
