+++ evento/appweb/templates/appweb/home.html
@@ -4,7 +4,7 @@
 <a href="{% url 'evento:cadastro_atividade' %}" class="top-menu">adicionar atividade aos meus eventos</a><br/>
 <a href="{% url 'tag:tag_usuario' %}" class="top-menu">adicionar tags a mim</a><br/>
 <a href="{% url 'tag:tag_evento' %}" class="top-menu">adicionar tags aos meus eventos</a><br/>
-<a href="{% url 'evento:cadastro_cupom' %}" class="top-menu">adicionar cupons aos meus eventos</a><br/>
+<a href="{% url 'cupom:cadastro_cupom' %}" class="top-menu">adicionar cupons aos meus eventos</a><br/>
 <a href="{% url 'evento:associar_evento' %}" class="top-menu">associar</a><br/>
 <a href="{% url 'inscricao:inscricao' %}" class="top-menu">inscrever-se</a><br/>
 <a href="{% url 'evento:criar_equipe' %}" class="top-menu">criar_equipe</a><br/>
