+++ evento/evento/urls.py
@@ -7,7 +7,6 @@ urlpatterns = [
 
 	url(r'^cadastroEvento/$',views.cadastroEvento.as_view(), name='cadastro_evento'),
 	url(r'^cadastroAtividade/$', views.CadastroAtividade.as_view(), name='cadastro_atividade'),
-	url(r'^cadastroCupom/$',views.cadastroCupom.as_view(), name='cadastro_cupom'),
 
 	url(r'^criarEquipe/$',views.criarEquipe.as_view(), name='criar_equipe'),
